refactor: keep existing type ownership
